构建结构有问题
