+++ run_set.json
@@ -3,7 +3,7 @@
   "DEFAULT": {
     "model_name": {
       "help": "model name",
-      "default": "unet",
+      "default": "unetr",
       "type": "str"
     },
     "load_run_id": {
